:sparkles: mv pkg/{runtime,}/signals
Continue moving stuff out of pkg/runtime.  the schemebuilder now lives
in pkg/scheme.
